+++ src/clblast_c.cpp
@@ -15,7 +15,7 @@
 
 #include "clblast_c.h"
 #include "clblast.h"
-#include "utilities.hpp"
+#include "utilities/utilities.hpp"
 
 // Shortcuts to the clblast namespace
 using float2 = clblast::float2;
