+++ src/tuning/kernels/transpose_fast.cpp
@@ -14,7 +14,7 @@
 #include <string>
 #include <vector>
 
-#include "utilities.hpp"
+#include "utilities/utilities.hpp"
 #include "tuning/tuning.hpp"
 
 namespace clblast {
