+++ src/tuning/kernels/xgemm_direct.cpp
@@ -16,7 +16,7 @@
 #include <string>
 #include <vector>
 
-#include "utilities.hpp"
+#include "utilities/utilities.hpp"
 #include "tuning/tuning.hpp"
 
 namespace clblast {
