+++ src/tuning/tuning.hpp
@@ -20,7 +20,7 @@
 
 #include <cltune.h>
 
-#include "utilities.hpp"
+#include "utilities/utilities.hpp"
 
 namespace clblast {
 // =================================================================================================
