+++ src/utilities/clblast_exceptions.cpp
@@ -13,7 +13,7 @@
 //
 // =================================================================================================
 
-#include "clblast_exceptions.hpp"
+#include "utilities/clblast_exceptions.hpp"
 
 namespace {
 // =================================================================================================
