+++ src/utilities/clblast_exceptions.hpp
