+++ src/utilities/utilities.cpp
@@ -11,7 +11,7 @@
 //
 // =================================================================================================
 
-#include "utilities.hpp"
+#include "utilities/utilities.hpp"
 
 #include <string>
 #include <vector>
