+++ test/performance/client.hpp
@@ -31,7 +31,7 @@
 #endif
 #include "clblast.h"
 
-#include "utilities.hpp"
+#include "utilities/utilities.hpp"
 
 namespace clblast {
 // =================================================================================================
